Update ann_practice.py
I have created a sample artificial neural network here with two hidden layers. I have used label encoder and standard scaler to transform the data. This data has features such as credit score, salary, age, gender etc and the target variable is whether the customer will leave the bank or not. So it evaluates the attrition rate of a bank and predicts weather a customer will leave the bank or not. 
I have used ReLU activation function to remove any linearity and sigmoid to give the binary output of TRUE or FALSE.