Add files via upload
A Classification artificial neural network that classifies whether a day will have high ozone exposure or not. It uses the dataset provided by the UCI website (https://archive.ics.uci.edu/ml/datasets/ozone+level+detection) which has 73 features for over 2000 records.
I have created an artificial neural network using the Keras library (Tensorflow backend) with two hidden layers which use rectifier activation and sigmoid activation for the output layer.
I was able to get an acc: 0.9980 and an accuracy of 0.9428007889546351 using the confusion matrix. 

The second part of the code creates a function that returns the classifier with an input of optimizer which we will use in this case 'adam'. Then we use the KerasClassifier from the Keras.wrappers.sci-kit_learn library , gridsearch. We get the best parameter and the best accuracy out of all the others. We use two batch sizes of 25 and 32, optimizer: 'adam', 'rmsprop'